% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{clean_list_data}
\alias{clean_list_data}
\title{Clean the NBA API data}
\usage{
clean_list_data(data)
}
\arguments{
\item{data}{object returned from `get_current_standings()`}
}
\description{
Clean the NBA API data
}
