% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{create_historical_table}
\alias{create_historical_table}
\title{Create a reactable from matching historical records}
\usage{
create_historical_table(data)
}
\arguments{
\item{data}{a records object of matching data}
}
\description{
Create a reactable from matching historical records
}
