% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{create_standings_table}
\alias{create_standings_table}
\title{Create a reactable for the app}
\usage{
create_standings_table(data, select = NULL)
}
\arguments{
\item{data}{either the eastern or western conference
standings tibble returned from `prepare_standings_data()`}

\item{select}{an argument passed to reactable. Default
is NULL. Other acceptible values are "single" and "multiple"}
}
\description{
Create a reactable for the app
}
