% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_current_standings}
\alias{get_current_standings}
\title{Get current NBA standings}
\usage{
get_current_standings(season = get_current_season_input())
}
\arguments{
\item{season}{a string in the format YYYY-YY (e.g. '2020-21'). 
default is the current season.}
}
\description{
Get current NBA standings
}
\details{
scrapes the nba.com/stats website for current
nba standings
}
