% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_historical_matching_records}
\alias{get_historical_matching_records}
\title{Get historical NBA matches}
\usage{
get_historical_matching_records(data, current_records)
}
\arguments{
\item{data}{all historical nba results}

\item{current_records}{the input record to match (e.g. '5-1')}
}
\description{
Get historical NBA matches
}
\details{
for a given input record (e.g. 3-4), 
extract all teams with a matching record in NBA history
}
