% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_matching_records}
\alias{get_matching_records}
\title{Get matching records}
\usage{
get_matching_records(
  data,
  input_wins,
  input_losses,
  input_plus_minus_min,
  input_plus_minus_max
)
}
\description{
Given inputs from the shiny application, extract
all matching historical records
}
