% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_record_summary}
\alias{get_record_summary}
\title{Summarize results for NBA W/L record}
\usage{
get_record_summary(data)
}
\arguments{
\item{data}{nba record data}
}
\value{
a tibble with summary records for each record
}
\description{
Summarize results for NBA W/L record
}
\details{
For a given NBA record, summarize the outcome for 
end of season results.
}
