% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{prepare_standings_data}
\alias{prepare_standings_data}
\title{Prepare standings data for tables}
\usage{
prepare_standings_data(records)
}
\description{
Prepare standings data for tables
}
\details{
scrapes data from nba.com/stats
creates two tibbles, one for the eastern conference
standings and western conference standings.
}
